pyproject.toml
[project]
name = "invenio-rdm-starter"
version = "13.0.8.1"
description = "Starter for the invenio-rdm turn-key research data management repository"
authors = [{ name = "Martin Fenner", email = "martin@front-matter.de" }]
readme = "README.md"
requires-python = ">=3.14,<3.15"
license = "MIT"
dependencies = [
    "invenio-app-rdm>=13.0.8,<14.0.0",
    "invenio-search[opensearch2]>=3.0.0",
    "invenio-s3>=3.0.2",
    "invenio-previewer>=3.3.0,<3.4.0",
    "invenio-oauthclient",
    "idutils",
    "setuptools>=70,<81",
    "gunicorn~=23.0.0",
    "structlog>=25.0",
    "prometheus-flask-exporter>=0.23.1",
]

[tool.uv.sources]
invenio-oauthclient = { git = "https://github.com/front-matter/invenio-oauthclient", branch = "maint-v5.x" }
idutils = { git = "https://github.com/front-matter/idutils", branch = "openalex" }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["site/invenio_rdm_starter/*.py"]

[tool.hatch.build.targets.wheel]
sources = { "site" = "" }

[tool.ruff]
line-length = 79
include = ["pyproject.toml", "site/**/*.py"]

[project.entry-points."invenio_base.blueprints"]
invenio_rdm_starter = "invenio_rdm_starter.views:create_blueprint"

[project.entry-points."invenio_base.api_blueprints"]
invenio_rdm_starter_api = "invenio_rdm_starter.views:create_api_blueprint"

[project.optional-dependencies]
docs = ["mkdocs-material>=9.7.0"]
